lab2 50%

+++ lab2/DebugExercise/DebugExercise1.java
@@ -5,9 +5,9 @@ package DebugExercise;
  * feature work.
  */
 public class DebugExercise1 {
-    public static int divideThenRound(int top, int bottom) {
-        int quotient = top / bottom;
-        int result = Math.round(quotient);
+    public static int divideThenRound(double top, double bottom) {
+        double quotient = top / bottom;
+        int result = (int) Math.round(quotient);
         return result;
     }
